"3:20pm. No, I can't. Actually the selector for the KL distribution will be the most complicated yet and I am going to have to make an entirely new kernel just for it.
I am literally going to have to do a triple block reduction with sequential reduction in the innermost dimension.

Wow.

Well, Spiral was made for things like this. I am glad I do not have to do it in C.

The reason why a triple reduction is necessary is because the tensor will have 4 dimensions: batch, action, support, probability.

Support would be just the multiplier for the probabilities denoting their decimal place, and the probabilities are just the softmax over over the activations that I plan to group in base 32.

My sense is that doing it like this will be a lot better than using binary, but I am not sure just yet. It will certainly be more stable than using logistic cross entropy.

These issues of data representation is something that I am going to have to go through and I might as well make this my first hurdle to clear.

Obviously I cannot use the one hot encoding for literally everything.

If I start playing large tournaments I want the ability to handle large inputs and rewards.

And given how quantile regression improved with multiple inputs, there is something to using distributional RL even with MC updates.

3:35pm. I do not feel like starting programming right now. Let me go get some exercise. Getting this done will take me a day or two and I want to take my time doing it properly."