Not bad. I quite like this.
This takes care of the block_reduce kernels.

Edit: Forgot to push this.

At any rate, I will stop here for the day as I am pretty tired. I've thought of factoring the functions in seq_broadcast, but it won't be that easy.

Rather than a triple reduction it is going to have to be a double reduction. There simply won't be enough threads for the third dimension so I'll let the grid iterate over it.

I've been thinking about these subtle issues.

...Nope, I do not have anymore energy. I want to factor the seq, but that damn thing is too tied up in the cub_block reduce. I am really not sure what I want to do with it.

I could go for a bit further, but I do not have enough to make a proper step.

I guess I'll take it easy and call it a day here. I usually do it at this time anyway.

This is not like back in 2016 and before. Back then I'd consider it a badge of honor to work till I drop, but now that I know that stress in me manifests as physical pain in the pit of my stomach I really am just fine with being lazy past 6pm.

5pm is the time when I do the clean up work.

I wish I could work 24h a day until I get to my goal instead of just getting maybe 4 hood hours in, but that is just a part and parcel of being a human.

Time goes on and my mind and body decays, but my power keeps growing.

Whether I can take it all back depends on whether I can understand the lesson of the last 15 years.

One day it won't be insignificant. And one day I will cause the Singularity.