Despite going to bed at 11pm yesterday, I got up like 7am today and have been watching this talk by Sutton since then.
http://videolectures.net/deeplearning2017_sutton_td_learning/

During the night it occurs to me that the reason logistic regression is so unstable to me is because I am not doing the gradients properly.

https://stats.stackexchange.com/questions/278771/how-is-the-cost-function-from-logistic-regression-derivated

Sure enough, it turns out the way I am doing it is too complex.

In the last library during my F# days, I fixed this issue by clipping the sigmoids, but now I could do better than that by fusing the activation with the cost much like in the softmax cross entropy.

So this is something I am going to do today.

After that I'll get to work on the new kernel. I think I will just extend dredo so it accepts reductions sequences.

Well forget that for a bit. Let me continue with the talk.