"9:40am. https://www.reddit.com/r/reinforcementlearning/comments/8hlmu8/lecture_richard_sutton_td_learning/?st=jgvwlntt&sh=b9c7355c
```
During the lecture it occured to me that there is one possibility that was missed so far in model based RL. Rather than do prediction with a function of the type `s * a -> s`, it would be much better to instead do it with a function of type `s -> a -> s`. The difference is that instead of using a function to make a point estimate of the next state given the current state and the selected action, you'd estimate a state for every possible action.

With that in place, it would be easy to do TD learning either on the states or on the internal features. It would be possible to make it even more efficient by predicting the internal state `i` using a function of type `s -> a -> s * i` and feeding that to the state value module.

This sort of formulation would open the door to also doing (internal) state prediction in the state value module and using the difference in predictions of the state of the two modules as a measure of intrinsic reward.

Raw prediction error in the world model is unsuitable as a measure of intrinsic reward because it cannot deal with noise, so probably the correct answer would be to use the error of a prediction of a prediction at some point.

Just my 2c. So far deep RL is split between PG and Q learning, and TD learning hasn't really entered the picture, but it is not really a crazy thing to consider with RNNs. It just hasn't been tried yet as far as I can tell.
```

10am. This talk really inspired me. Last time it was natural gradients, and now it is TD learning. I did not expect this, it just goes to show that you never know when inspiration is going to strike.

I can sense it now. My programmer's intuition is telling me that the pieces fit. I can make use of this idea together with RNNs. I do in fact have an alternative to Q learning, PG and MC updates. Truly brilliant.

I kind of feel like going back to Littman and Isbel's course and covering the basics again.

Sutton's arguments in favor of TD over MC are convincing to me even from a gradient propagation perspective. It is in fact what I should be doing. I need to make TD learning work with NNs. Before when I looked at it, I just thought 'where is the action' and dismissed it."