I was amazingly useless today, but I think I finally understand what TD updates are all about. In the lecture by Sutton it all somehow just clicked for me.
I think I understand the deadly triad. Either way, I'll be fine with NNs as long as I do not do bootstraping. And if I do bootstrapping, I need to make sure to push the updates in the direction of the natural gradient.

I am not sure whether differentiable plasticity or Reptile or EigenNets will be the key to cracking this, but the answer will be there somewhere.

I was really going to do programming, but apart from the talk, I was just in a frenzy to take in information. Considering the talk was only 1.5 hours, I am not sure what took so much of my time today. I read the RL book by him and watched the lectures, but could not get much out of it.

Well that is usually the case. Unless you are focused on a task, the time will vanish in the blink of an eye.

I really wish I got something more out of the day other than the talk. If I could have looked into the future, I would have switched to programming 4.5 hours ago, but now my stress is near an extreme so I am stopping here.

Damn it, I was really looking for an alternative and I saw a similar talk by him years ago, but back then I did not really have the same feel into it.

Now that I am actually going at it with the intent to make specific implementation and that I have cleared the burden of making a language, my mind is laser focused on all the tiny details.

Sutton is right about everything. The view he has essentailly deals with all of the difficulties of doing MC updates.

Cracking bootstrapped predictions with RNNs will absolutely have to be done to get to tier one. Instead of doing maximum likelyhood estimates I absolutely need to be doing certainty equivalence estimates.

I will whatever I can on my side. I will definitely experiment with TD learning as MC methods will not scale.

Tomorrow I will deal with the kernel without fail. I do not think I'll be distracted with anything else.

I do not think there is much else past this point anyway.

I think I've covered the essentials at this point.