"7:45am. I got up 10m ago. In contrast to yesterday, I feel fully fresh now. Forgot to read Overlord.
https://boards.4chan.org/a/thread/172600399/jahysama-wa-kujikenai-jahysama-wont-be

7:50am. Ah, screw it let me chill a bit. Let me start off with a commit though.

I won't work on the double reduce right away. The TD stuff shook my resolve a little. I'll focus only on a single distribution for now and won't move until I confirm it works better. Realistically, until the reward scaling issues are resolved in standard SGD, I can't really go to TD learning.

Getting natural gradient methods to work is key to making bootstrapping work.

https://arxiv.org/abs/1602.07714
Learning values across many orders of magnitude

Ohhh, this seems interesting.

https://arxiv.org/abs/1712.02390
Noisy Natural Gradient as Variational Inference

Let me catch up to Overlord and then I will read these papers. After that I will start programming.

What a pain in the ass. There is really not a simple way to abstract adding a whole extra dimension to the algorithm. I am going to have to copy paste it. Adding that ddef to the seq was a mistake. I am going to have to remove it.

8:20am. Now intermission is left, but let me read the papers.

8:35am. The rescaling in the paper is quite complicated.

8:40am. Bayes by Backprop?

That seems like an interesting paper.

https://arxiv.org/abs/1505.05424
Weight Uncertainty in Neural Networks

This is that paper.

8:55am. Ok, I am done with the paper. Let me read the latest chapter and then I will get cracking.

Even though I am convinced of the merits of Sutton's argument regarding TD learning, there is really no point in going for it when I can't even do something as simple as scaling the rewards properly. It might be worth trying with QR, but...what a pain.

I really have a lot of things I need to try. I'll give it a try with proper QR. The way I am using it for MC updates is not what it was intended for anyway.

Let me just get standard SL to work on poker. Once I form the base, I can pick out the methods as they come to me at my leisure.

The `Learning values across many orders of magnitude` is really too complicated to be considered a solution.

9:10am. I am ready to start. Damn that `ddef`, I should not have moved to that pattern. I am going to have to start by undoing it.

Then I will just copy paste the seq kernel and adapt it so it takes in one more dimension."

This is how my mornings go usually. If I do not mix play and work, what happens is that like yesterday when I obsessively studied my stress goes through the roof. I'll take it easy starting from here.