This takes care of the new kernel.
Well, not quite. I need to write the unit test for it. Let me just do it. After that I will bring it into action.

I also did some refactoring of the other kernel so it is not so abominable. As it turns out, I pretty much forgot why ddef existed since it did multiple things.

Doing the KL thing won't be difficult. Let me just do it. I really need to catch up to the state of the art.