Somehow despite me finishing this, the way I've held myself during the duration of the session is demoralizing. I've gotten irate over my perfectionism which just made things worse. At this point, I just want to go forward and make solid progress. I do not want to get bogged down in refactorings as I have for the past years.
I want the power to edit myself. But more than that, if I cannot even deal with these simple NNs, how could I possibly handle as something as complex as the brain? These NNs that I will be working with are the basis of everything.

I keep worrying about various aspects of the algorithm and futilely trying to improve them with my vastly insufficient reasoning abilities.

If I am going to agonize, let me do it after I reach the state of the art. Let me clear away all of the burdens so I can tackle the problem with a clear mind.

https://www.youtube.com/watch?v=N5oZIO8pE40
From Generative Models to Generative Agents - Koray Kavukcuoglu

Let me watch this video while I take a break. I'll get to work on the selector after that.

The state of the art. That is the one singular desire I need to focus on if I want to accomplish something of value by the time of the next update. I might or might not accomplish anything once I reach it, but I have to learn to walk before I run, that is the most basic of the basics.