I got it to work, but so far literally nothing works better than a single linear layer + square error.
How could this be happening?

I am getting quite worse results with scaling. Going from 1 to 2 layers with nonlinearities does slow it down.

So probably the real reason why square error works better is because the lower number of weights makes it generalize better.

The more powerful models I am using just make it overfit.

I have two choices at this juncture:

1) Move to using RNNs. From here move to implementing differentiable plasticity.

2) Implement the replay buffer and move to minibatch learning. From there move to implementing Reptile.

Of course, I need to do all of that, but which one should I do first?

...I am going to pick the third option.

Implement the replay buffer, but use it to do online learning. Reptile as it is right now is not quite ready for minibatch learning. I'd have to split the weights and do the same things as in dynamic evaluation paper to get it going.

I think I will go with that.

Really, if this is the best NNs can do then I am definitely screwed. Reptile is simply going to come through and RNNs are going to have to add ontop of that.

Really, Reptile sure picked a good time to get released. It was on April 1st right after I finished LN for the first time.

This timing is heaven sent, so I migth as well put my faith in it.

I'll continue with the plan of trying out things for a few months. If both Reptile and RNNs cannot get this thing to work really well, then most likely I am going to have to wait for more stable models. In that case, I'll change gears and maybe use Spiral to make games or something like that.

Let me take a break here. I did some good work here.