"10:40am. I am still thinking about it.
Basically it is not as easy as just getting RNNs to work.

In order to get them to work or the online Reptile, I really need to do everything would need to also make the replay buffer work.

The real problem is the game itself. I sort of have the trace separate of the player which makes everything that much more difficult.

Basically I need to redo everything from the beginning. It won't be literally everything, but getting this done will require a change of mindset. Rather than do an extension, I need to make the goal work and then refit the rest around it.

10:45am. Let me just have breakfast here. I am still rolling it in my head.

It seems that I will get the online version of the replay buffer, RNNs and online Reptile all done in one swoop. Having that will also make it easy to switch between MC and bootstrapped updates.

I might as well go for it. It has been a while since I did a major rewrite."