I've thinking about it for the past 1.5 hour. I haven't written a line of code.
Honestly, I am completely at loss how to do this. The complicated ways the data would flow through the system with the replay buffer is making me doubt what sort of design would be right here. I understand the individual pieces, but not how they should fit together.

Let me step out for a bit.

It will take a me a while to work through this mental block. Design is usually the hardest part of programming.