Well, aside from chilling which I am going to do before programming, I did a lot of thinking how I am going to do the the true QR cost function with predictions errors. Not only have I figured it out, I actually made a significant conceptual breakthrough regarding iteration.
For the first time I can clearly visualize that `inl k j i -> ...` does not necessarily mean that `i` must be the innermost dimension. In the 2d reduce I made `inl i j -> ...` where `i` was actually the innermost dimension despite being on the outside. At that time I took advantage of it, but I never had a direct fragment that I could use to do reasoning. Now I can see it.

And actually this bit of insight is the crucial part of doing the matrix multiply. I will be able to do it without much trouble now.

Brilliant, truly brilliant.

Now, let me chill for a bit and then I will start programming.

I have absolutely all the pieces that I need now. I understand how to extend the trace thanks to all the thought I put into it yesterday. It will take me a while to get it all done, but by the time of the next update I might have the state of the art player done. The path in my mind is clear.