I got it to work. It is pretty messy, but with this I think I will call it a day. I had enough. This work is tedious and I do not feel like pushing myself to clear it out.
Tomorrow I will clean this up and get the rest of the tests to work.

After that I will put in Reptile into `trace.process` and replace the feedforward layers with MIRNNs. QR, Reptile and differentiable plasticity are basically the 3 biggest things in RL in the last 3 years. All 3 came out in the last six months.

If they do not give me the lift I need to reach a player good enough to a player good enough to beat microstakes then I can just pack my bags and move somewhere else. A part of why I spent all this time working on the language and the libraries is so the field has time to set itself up.

Assuming I can get MC updates to work better, I will try SARSA or DQL as well. After that I will move to a model based approach and move to TD learning.

There is also the paper by geri that I am looking forward to reading. It could be huge and it could be nothing. I am definitely curious as to how it will turn out either way.