Fixed a little typo that I spotted in addition to what I got in the PR.
Also I noticed he put in the path for the Spiral dll in Debug. Release is really much better.

This should be good for starting things off for the day. Today on the agenda is to clean up the stuff from yesterday, replace the feedforward layers with MI-RNNs and then implement Reptile.

Doing all of the above won't be hard, but once I have Reptile I am not sure what to do about it. MAML and it are intended as pretraining procedures.

If I skip the fine tuning with SGD step, will it still work well? Only one way to find out. I've looked for further benchmarks on Reptile, but I haven't been able to find anything. It is just too fresh at this point. Maybe some stuff will come out after the NIPS deadline.