I figured out what the problem was. It was a dumb copy paste error in the test itself. Damn the rescaling.
Actually, rather than a lengthy debugging, I was spared the process by my intution. I got the sense that mostly likely the error is in the test itself and sure enough it was.

Before I move on to RNNs, I need to investigate why the code bloated up. In the previous version it was 5.2k and here it is 6.5k.

What could this possibly be?