I managed to trim it a little, but I can't find the source of the 600 LOC shortfal. This is as about as much effort as I feel putting into it.
What a pain in the ass. I am so annoyed by this.

I am thinking it might be because of the list function. At any rate, something like this is not worth wracking my brain over anymore. I could probably get down to the exact root of it, but I do not feel spending that much time on this.

Enough is enough.

I think I spent like 3h on debugging this crap. The errors always turn out to be something trivial that I could have avoided had my memory been better.

This is about what I suspected anyway.

Now is time for the more fun stuff. Let me try out an RNN with this.

I'll replace the feedforward LN layer with a miln recurrent layer.

It is a one line change that will work fine if I got all of it right.