"2:30pm. That covers RNNs. They weren't hard to try out.
So far the tally is linear > feedforward > recurrent. This is is roughly in the opposite order of their power. By the looks of it the greater expressive power of recurrent and feedforward models only leads to more over-fitting on the task I am trying them out on.

In RL contexts that means markedly subpar performance. I couldn't possibly have these agents handle bets in my stead.

If this was 2016, my options would be severely limited at this point. But let me give Reptile and differentiable plasticity a shot. If they can make MC updates work better then that would open the doors to bootstrapping methods.

If not, I might have to spend time looking into natural gradient methods. There would be simply no other choice in that case - there is absolutely no reason to use NNs if they can't even beat linear regression. I don't think I'll have to resort to this though.

2:45pm. I do not feel like diving into Reptile right now. Let me go outside for a bit. I need a breather.

It won't be hard to do at any rate. "