"4:30pm. Done with chores. When I came back I felt like lying down and just thinking for a while.
To be honest, I do not feel like implementing Reptile just now. Assuming it works which I believe it will - I know how to solve the issue of needing target networks. Having meta-optimization puts on the table an unique opportunity - I can freeze the targets on the subsequent gradient steps.

By that I mean make the regular bootstrapping update on the first step and on the subsequent steps reuse the targets from the first step. Of course at the end make the meta-update.

What I like about this is that it gives a firm negative answer to the question of whether gradients should be propagated through the targets. Here the answer is no.

4:45pm. Yeah, I am done for the day. No way am I starting now. I'll build up my motivation for tomorrow. Thinking up the ideas like the above really motivate me to keep going. They are mental equivalent of money I layer convert into code.

I do believe in the above idea. It is either something like the above or natural gradient related methods which will get me to the end.

Benchmarking State-of-the-Art Deep Learning Software Tools
https://arxiv.org/pdf/1608.07249.pdf

Let me read this and then I will be done.

https://www.reddit.com/r/programming/comments/8ict1m/herb_sutter_at_build_2018_how_to_adopt_modern_c17/?st=jh231cc0&sh=0c27a4d7

There was also a video by Herb Sutter on some poker thing which piqued my interest."