Basically it does not work even on things that it should.
Far from Spiral making this easy, all the noise from the data transformations is making this quite difficult.

And then main thing that has changed is that I am no longer optimizing after every game, but after every hand. I need to make sure that this is not what is making it diverge. If it is, then things will be quite difficult.