"2:45pm.
```
Starting timing for: iteration 0
Winrate is 659 and 341 out of 1000.
The time was 00:00:07.2994915 for: iteration 0
Starting timing for: iteration 1
Winrate is 679 and 321 out of 1000.
The time was 00:00:09.5386472 for: iteration 1
Starting timing for: iteration 2
Winrate is 704 and 296 out of 1000.
The time was 00:00:10.9940683 for: iteration 2
Starting timing for: iteration 3
Winrate is 741 and 259 out of 1000.
The time was 00:00:10.8668233 for: iteration 3
Starting timing for: iteration 4
Winrate is 666 and 334 out of 1000.
The time was 00:00:10.0224357 for: iteration 4
Starting timing for: iteration 5
Winrate is 707 and 293 out of 1000.
The time was 00:00:11.9969956 for: iteration 5
Starting timing for: iteration 6
Winrate is 704 and 296 out of 1000.
The time was 00:00:10.6288275 for: iteration 6
Starting timing for: iteration 7
Winrate is 681 and 319 out of 1000.
The time was 00:00:10.9332358 for: iteration 7
Starting timing for: iteration 8
Winrate is 588 and 412 out of 1000.
The time was 00:00:11.4514759 for: iteration 8
Starting timing for: iteration 9
Winrate is 449 and 551 out of 1000.
The time was 00:00:12.4770139 for: iteration 9
```

This is on the RNN. Horrible.

2:55pm. I am thinking right now.

...I do not think there is an error in my code, but rather in my understanding of what Reptile is supposed to be doing. Of course it is not going to be that easy. I need to look at the places where Reptile works a bit more carefully. Diving by the learning rate makes it equivalent to 1. There is no way that is right. It is no wonder it blows up.

I guess my main problem now is that Reptile is supposed to be used for multi task learning and here I am not really doing that, but using it in the hopes it will improve optimization on a single task.

One thing that is striking is just how much going from optimizing on a single game to optimizing a single hand made things worse. Maybe the secret to making things work is to optimize on really long sequences? I am guessing the Baird's counter-example could be made better if it was optimized in batch mode.

I'll take a break from programming here in order to study the online implementation of Reptile by the author."