3:25pm.
```
Starting timing for: iteration 0
Winrate is 645 and 355 out of 1000.
The time was 00:00:07.1069181 for: iteration 0
Starting timing for: iteration 1
Winrate is 692 and 308 out of 1000.
The time was 00:00:07.8347729 for: iteration 1
Starting timing for: iteration 2
Winrate is 734 and 266 out of 1000.
The time was 00:00:09.6901596 for: iteration 2
Starting timing for: iteration 3
Winrate is 761 and 239 out of 1000.
The time was 00:00:11.0727680 for: iteration 3
Starting timing for: iteration 4
Winrate is 773 and 227 out of 1000.
The time was 00:00:10.0021227 for: iteration 4
Starting timing for: iteration 5
Winrate is 765 and 235 out of 1000.
The time was 00:00:11.1432205 for: iteration 5
Starting timing for: iteration 6
Winrate is 762 and 238 out of 1000.
The time was 00:00:11.2915751 for: iteration 6
Starting timing for: iteration 7
Winrate is 751 and 249 out of 1000.
The time was 00:00:12.1424866 for: iteration 7
Starting timing for: iteration 8
Winrate is 777 and 223 out of 1000.
The time was 00:00:12.2860978 for: iteration 8
Starting timing for: iteration 9
Winrate is 772 and 228 out of 1000.
The time was 00:00:12.4755385 for: iteration 9
```

This is with meta_lr of 0.3 and the inner_lr of 0.003. But really, what if I set the meta_lr to 1?

3:30pm.

```
Starting timing for: iteration 0
Winrate is 654 and 346 out of 1000.
The time was 00:00:15.5612829 for: iteration 0
Starting timing for: iteration 1
Winrate is 736 and 264 out of 1000.
The time was 00:00:15.8987283 for: iteration 1
Starting timing for: iteration 2
Winrate is 775 and 225 out of 1000.
The time was 00:00:14.3632598 for: iteration 2
Starting timing for: iteration 3
Winrate is 740 and 260 out of 1000.
The time was 00:00:16.3293144 for: iteration 3
Starting timing for: iteration 4
Winrate is 752 and 248 out of 1000.
The time was 00:00:15.5016083 for: iteration 4
Starting timing for: iteration 5
Winrate is 740 and 260 out of 1000.
The time was 00:00:15.9539344 for: iteration 5
Starting timing for: iteration 6
Winrate is 781 and 219 out of 1000.
The time was 00:00:15.9199922 for: iteration 6
Starting timing for: iteration 7
Winrate is 764 and 236 out of 1000.
The time was 00:00:15.9512157 for: iteration 7
Starting timing for: iteration 8
Winrate is 769 and 231 out of 1000.
The time was 00:00:15.5117902 for: iteration 8
Starting timing for: iteration 9
Winrate is 742 and 258 out of 1000.
The time was 00:00:13.9211116 for: iteration 9
```

Then I get this. Damn, a single iteration really takes a while now.

Let me compare it with a normal run.

```
Starting timing for: iteration 0
Winrate is 675 and 325 out of 1000.
The time was 00:00:04.6496046 for: iteration 0
Starting timing for: iteration 1
Winrate is 714 and 286 out of 1000.
The time was 00:00:04.9938658 for: iteration 1
Starting timing for: iteration 2
Winrate is 745 and 255 out of 1000.
The time was 00:00:04.9705424 for: iteration 2
Starting timing for: iteration 3
Winrate is 765 and 235 out of 1000.
The time was 00:00:05.0403464 for: iteration 3
Starting timing for: iteration 4
Winrate is 749 and 251 out of 1000.
The time was 00:00:04.9875465 for: iteration 4
Starting timing for: iteration 5
Winrate is 747 and 253 out of 1000.
The time was 00:00:04.7457833 for: iteration 5
Starting timing for: iteration 6
Winrate is 769 and 231 out of 1000.
The time was 00:00:04.9692537 for: iteration 6
Starting timing for: iteration 7
Winrate is 745 and 255 out of 1000.
The time was 00:00:04.8325493 for: iteration 7
Starting timing for: iteration 8
Winrate is 759 and 241 out of 1000.
The time was 00:00:04.9126061 for: iteration 8
Starting timing for: iteration 9
Winrate is 748 and 252 out of 1000.
The time was 00:00:04.8066445 for: iteration 9
```

I am really getting nothing from Reptile.

Let me stop here just for a bit. I'll keep it in mind for a while, but I won't bother trying to find uses for this for the time being.

The next thing on the list of things to implement would be differentiable plasticity and bootstrapping methods in general.

But before that, let me read that paper on differentiable programming. I also want to revisit some papers.