Ok, enough of this. I've been woke up for an hour already. It is time that I put some focus into how I am going to do deep Q learning.
As per plan I will get rid of the deep MC stuff as it is too unstable. It would only get worse with longer sequences. I'll also get rid of the Reptile player.

Let me focus on the goal for a bit and then I will start. Just like with Reptile, I will have to come up with a concrete plan.