Since my focus is starting to slip, how about I take a break here?
I really did quite a bit of work in preparation for this while I was implementing Reptile.

As far as the square error is concerned I think this will be enough.

Well, maybe later I will have to extend it with the target network, but forget that for now.

Right now I have to make sure that it is not optimizing per hand...

Though actually, it should be fine either way.

Let me do it per hand just so I can compare the stability of the two methods. Or at least per game since I removed it...no, I'll put it back in and try it properly.

After that I will put in QR. Only after that will I add the rest of the tricks.

Edit: Forgot to push this.

Time to resume this thing. I will get this thing working today.

Q learning is something I will be working for a while and today will be the first step.