"2:15pm. I am thinking right now.
I am completely confused by this. Basically, relu units cause the feedforward net to always blow up. Tanh work if you can call this working.

I really regret it that I cannot answer the question of why one works and the other doesn't. Can this really be called expertise?

This is not working. At this point deep Q learning is not even on the table. It does not matter one bit if it works on Atari when it works this poorly here.

2:20pm. Before I throw in the towel I need to put in a little more effort. The relu vs tanh thing shows that architectures make a lot of difference to stability of the entire thing.

So in order to stabilize this, I will go to per game optimization. Then I will put in quantile regression. Then I will try out differentiable plasticity.

If that does not work, then I can either go back to MC updates or do something else entirely."