In the last commit I had to run because of thunder, but now that it has passed let me go into a bit more detail.
Basically, deep Q learning, at least without target networks is a complete failure. And with target networks I just know it will be 100x slower than its full potential.

The thought of doing target networks with RNNs just makes me want to hurl.

Honestly, QR works so poorly that I am not sure whether I implemented it correctly, but works on par with squared error with a distribution size or two or three so I'll give it the benefit of the doubt. I did put it through grad checking a while back so I'll just roll with it.

Before I go back to MC updates, I will implement differentiable plasticity and plug that RNN as the work horse.

I simply have to check whether DP as an architecture is anything special with regard to RL. If it can rein in this wild instability, then I will know that I am going in the right direction.

I doubt that DP will really make that much of a difference. It does not feel that it is the architecture that will make the SGD updates be identical to the natural gradient ones, but I feel that I should check.

Bootstrapped RL is the one true test of neural network architectures. If NNs can chew through it without target networks, then it will be obvious that RL is ready for prime time.

At the very least I will end up replacing M-IRNNs and LSTMs with it so I should do it anyway.

With this I am down to the last stretch.

Once I do DP, apart from the replay buffer I will have gone through the list.

Hopefully after the NIPS paper submission deadline passes some good things will come out.