I wasted most of the day, but the lecture by Ankit Pater made it all up.
https://www.reddit.com/r/MachineLearning/comments/8jduf8/r_convnets_from_first_principles_ankit_patel

"I found this talk quite fascinating and way beyond what I expected given the subject matter. I expected it to be dry, but it was definitely worth watching. Hopefully the rest of the lectures will be on ball.

As a side note, [the videos](https://www.researchgate.net/project/Theories-of-Deep-Learning) from [the course](https://stats385.github.io/) are placed in a really awkward place so I barely spotted them on the main page."

I am not kidding. I am really thrilled by this talk.

If I could understand the probabilistic interpretations of NNs I could understand what the architectures are actually optimizing for and what the components are actually doing.

With that understanding I could predict why bootstrapping DQN is failing and what do I need to do to fix this.

I am going to go through the rest of the lectures from this course and then I am going to look up the guy's papers. After that maybe it will be time to revisit the PGM lectures and learn some proper math.

After that I'll give RL a shot again.

Right now I am basically just wasting my time.

At this point, building up the foundations is no longer an issue. I have Spiral, I have the ML library, I have the game. I have my programming skills. I am good in this area.

What I do not have is the understanding necessary to create a good RL agent. I simply do not have the tools to reason about what I am doing in this area. It is not enough for me.

If I had the necessary understanding, I could create a top notch agent and the interface for in maybe a month or two of work.

I am inspired. I'll abandom my original plan of trying things out and learn Bayesian stuff.

In 2015 I went through the PGM course and hated the experience, but I think it is about time I revisit all of this. I am not the same as 2.5 years ago.

Back then it sort of felt like that stuff was incredibly complicated, but it is clicking for me bit by bit.

A lot of the math stuff is out of my league right now, but this seems like a good time shore everything up.

Completing my education will not take me a short amount of time. It might take me years, but I am prepared. I mean I spent a year making a library and then 1.5 years making a language. Trying to attain perfection is par the course for me.

But if I can make this step successfully, I will be able to to close the gap between me and the guys at Deepmind who are my competitors in this race for the Singularity.

There is no doubt that with my programming skill and a deep understanding of machine learning, I would be quite a terror to behold.

Now I am quite well suited to get some shitty corporate job.

My new goal is this - study until the understanding is attained.

Unlike before, I think that this time this will be accompanied by programming, but instead of doing deep learning, maybe I will give making a probabilistic programming library in Spiral a try. I will see as I go along.

It is unlikely I will just be studying math - this intersection of Bayesian methods and deep learning is really at the heart of everything I would want from both schools of thought.

On its own deep learning is completely opaque and on its own Bayesian methods are completely unscalable.

This here is a vision of brighter tomorrow - it is the point where I need to press down on and push until it yields and opens.