The other talks are just so boring. I am sick of them. It seems the one by Ankit Patel will be the only one really worth watching.
https://arxiv.org/abs/1805.00909
Reinforcement Learning and Control as Probabilistic Inference - Tutorial and Review

This was posted on the ML sub recently so let me read it and I will call it a day.

I think a decent prediction of how things are going to go is that NN will get a Bayesian interpretation which will allow them to be used within the context of the framework.

I'll ditch NNs for a bit and eventually figure out how to bring them back in.

I am thinking that solving the poker minigame using Bayesian methods would be a fruitful approach towards starting on the path.