Yeah, I think I will avoid anything deep when doing RL. How about I just make the most souped up model with differentiable plasticity and all the bells and whistles and then use a linear controller on top? That should get me what I want with only the minimum of hassle.
I can't deal with this shit.

I mean is it really that difficult to stabilize a net with just a single hidden layer?

It probably is. I can't think of a single damn thing here.

I'll read the papers by Ankit tomorrow. I am done for the day. I really can't go on anymore.