I am up. I had a rough night. I was assaulted by insight and have managed to solve pretty much everything. I realized something truly fundamental about differentiable programming and I understand how to form associations between actions and states now.
I was not sure whether I should have projected the action to a state and multiply or whether I should take the action and feed it as part of the input. It turns out both are wrong. In order to form an association I need to project it to a vector of zeroes and then add it to the state. That would essentially create a channel that cannot be ignored between state and action leading into it.

All the pieces are finally together. TD/Q learning is simply too difficult - saying that natural gradient updates will be needed for it is just hoping for a miracle at this point. And MC updates are not scalable just as Sutton says.

I had the idea of propagating rewards through the recurrent dynamics, and the only way to do that is policy gradient. Its great flaw of it essentially being mutable global state means I can use it to accumulate rewards over time horizons one at a time without having to do MC updates at the top level.

The action-input direct association and recurrent channels will ensure that the paths that lead to the high reward states get strengthened. And the style of updates I am doing will ensure that I do not run into the issues of either MC or bootstrapped updates.

As a added bonus I could even do model based learning using this method without having to block the gradients from the controller. In fact I realized that while the model cannot optimize for sums of rewards, it can predict them just fine on an individual basis. For tougher things I could use this to do imaginary training.

Doing it like I have in mind is literally doing it in the simplest possible way. I've distilled deep RL down to its very essence.

This is of course not nearly enough to punch through the box, but for poker it should do just fine.